changed node selection after delete
git-svn-id: svn://136.177.114.72/svn_GW/WPhast/trunk@2102 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/ZoneActor.cpp
@@ -141,7 +141,15 @@ void CZoneActor::Remove(CPropertyTreeControlBar* pTreeControlBar)
 			this->m_dwPrevSiblingItemData = pTreeCtrl->GetItemData(this->m_hInsertAfter);
 		}
 		ASSERT(this->m_hParent && this->m_hInsertAfter);
-		if (this->m_hParent) pTreeCtrl->SelectItem(this->m_hParent);
+		HTREEITEM next;
+		if (next = pTreeCtrl->GetNextSiblingItem(this->m_hti))
+		{
+			pTreeCtrl->SelectItem(next);
+		}
+		else 
+		{
+			if (this->m_hParent) pTreeCtrl->SelectItem(this->m_hParent);
+		}
 		VERIFY(pTreeCtrl->DeleteItem(this->m_hti));
 	}
 }
